I've hit another blocking point, in that D's spawn will not let me launch a class method as a thread. I get this error:
[rstovall@gabbie thread_pool]$ dmd -c thread_pool.d 
thread_pool.d(44): Error: template std.concurrency.spawn does not match any function template declaration
/usr/include/d/dmd/phobos/std/concurrency.d(401): Error: template std.concurrency.spawn(T...) cannot deduce template function from argument types !()(void delegate(),Tid)

This is not as bad as it seems, because D's scoping model at the module level can provide the needed encapsulation, but it will require another rethink of the design. I'm saving this as a reference point; it does not work, of course.