nrtb::confreader boost cleanup complete with behavior changes.
The boost lexical_throws when a requested conversion can't be
accomplished, while stringstream does not. We were using this
functionality to filter in some cases. This was an arbitrary
decision which does not affect most use cases, so we did not
attempt to replicate that functionality. Instead you get the
normal iostream behavior: "duh" would map to 0 for numerics,
and "71.2723" would map to 71 for ints.

Also, to keep it simple on the first whitespace delimited value
after the name is returned by the get and get all transformative
templates. If you want the full value, it's still available by
accessing the internal list.