sim_core::get_obj_copies complete; very dangerous!
Task @T14 : I completed this method because it was originally spec'd
but in hind-sight it is impossible to implement without exposing
too much of the world's internals to possible external manipulation.
I'm going to leave it in for alpha, but we need to seriously consider
removing this when we start beta level code.