Checkpoint while developing sim_core unit test.
Task @T18 : Add and remove object while not running is working
properly. run_sim() is starting and time tracking is spot on, but
start_sim() does not seem to be returning properly (or at least
nothing else is moving after it's call. CPU is not bound.. in fact
utilization is very low at 50hz.