Set up the template for the diff_steer effector.
Task @T29 : diff_steer is a pre-effector which implements a rudimentary
diff steering (i.e. turtle-bots or bulldozers) drive system. This was
picked because it's the easiest contact drive to create and because it
is sufficent to the needs of the NRTB alpha.
This compiles but does nothing yet; it's just a configured effector
template.