Revised this module to include three classes.
Task @T29 : After design review, this module needs three objects,
one pre-effector, one post-effector, and a master which creates
them, attaches them to an object, and provides a control interface.
All of this has been roughed out and the really trivial methods
have been created. This code compiles but does nothing as the
heavy lifting remains to be done.