Completed constructors for the diff_steer class.
Task @T29 : The constructor chain is complete. I'm not sure I'm
handling the shared pointers for the effectors in a safe way, so
I put in a test for that in the diff_steer constructor. If not, it'll
blow on construction and I'll need to add new add_pre and add_post
specializations to base_object to accomidate this use case.