Fixed a library ordering issue which was causing link errors.
Task @T52 : I had to reverse the order of sim_engine.a and
nrtb_common.a to make an evil linking issue go away. I'm not
sure why it should matter, but my opinion matters less than that
of the complier. We should be ready to work on the unit test now.