Cleaned up message handling in bot_mk1.
Task @T70 : While playing with controlling a bot by connecting
to simengine using telnet, it was discovered that null messages
could segfault the bot message handler, and that the reciever
was unnecessarily sensitive to line endings. A coouple of simple
changes made bot_mk1 (and therefore the system) much more robust.