Added the ability to suppress writing of empty quantas.
Task @T71 : During review, it became obvious that writing sim
output when there are no objects present is a waste of time and
space. I added a new configuration parameter, "write_zeros". When
true (1), all simulation quanta are logged; when false (0) writing
is suppressed when there are no objects present in the simulation.
Tested and ready to go, including configuraiton file updates.