Completed the fix for bot_mk1 braking issue.
Bug @B11 : Turned out the issue was due to an overly agressive value
for braking. Reduced the bot_mk1's max braking value from 1e6 to 2e5
(twice the drive value) and the problem went away. I'm good with this
fix for alpha, but we may need to revisit the math in diff_steer later
to figure out where we were losing it.
Unit tested and ready to go, this completes coding and unit test for
bug 11.