First part of cal_bot complete and working.
Task @T74 : Work on the first bot control program, cal_bot, is
progressing nicely. At this point it compiles, connects to the
simengine and runs a one second (configurable) accelleration test
to determine the actual accelleration in m/s for the bot_mk1.
Still remaining are sections to determine the actual braking rate
and turn rate.