More revisions to panopticon.
Task @T60 : Reverted the changes from the previous commit and altered
panopticon so that the add mathod takes a base_object and creates
the sensor_rec in the list context. This both cleans up the code in
sim_core::run_sim and makes experimentation easier. The problem we've
been chasing may still remain.