Checkpoint on radar_mk1 development.
Tasks @T61R @T62 : Made the returned target azimuth relative to the
parent object. Added a target object to the unit test rotated so
that it points directly towards the other two objects, returned an
azimuth within 0.001% of 0.. close enough for government work. This
completes work on the radar_mk1 class. Need to formalize the checks
in the unit test and then we'll be done with it as well.