Checkpoint on the rail_gun_mk1 development.
Task @T84 : Started outlining the class header. Most of the
variables are defined, and it's been decided that this class will
be tied to the bot_mk1 to have access to the talk-back queue. This
is needed to allow the rail_gun to notify the BCP when aiming,
charging or auto-fire is accomplished.