Created interfaces and started on unit test.
Defined three abstract classes which define interfaces for modules
to use communicating with bots. They are "bcp_sender",
"cmd_interface", and "tickable". bcp_sender provides a method to
allow the module to send a message to the BCP; cmd_interface allows
the module to present a commmand as if it came from the BCP, and
tickable allows a module to plug into a bot's scheduling system.
All are pure virtual. Working on the the unit test now, which will
demonstrate typical use scenarios.