Completed unit bot_interfaces unit test.
As this module only defines abstract interface classes, the only
testing consists of making sure the interfaces are usuable and
provide the functionality promised. To that end this unit test
implements each of the abstract classes and uses them for simple
tasks; a failure would be a compile failure.  This builds,
compiles and runs as expected. This module is ready to merge.