Added a bit of process control to bot_mk1::wait_for_tick()
Task @T86 : Realized I had not dealt with the dying bot case in
wait_for_tick(). I'm now testing for alive at entry and exit
(there's a pause in there) and throwing if the bot is dead or
dying. Also made sure that the destructor frobs the control variable
to notify all tasks on the way out.