Better take on the to/from polar conversion. Not quite right yet.
Tasks @T92 @T93 @T94 : Did some research, much better. Polor encoding
was a bit off, cartesian was WAY off. Better, but needs work to deal
with corner cases. It does not deal with with negatatives, and there
is one cartesian case (1,1,1) which is not converting back properly.
Significantly more work to be done.