The collision test works with moving objects, but the game cycle needs work.
Task #26 : Completed unit test changes which show the collision test is
working well for mobile v. mobile collisions. The problem is the current
test is forward looking (i.e. it checks from current position forward),
while the current game cycle moves the object before the check is made.
Checking for collision before movement commits is a good thing, so we
will need to alter base_object and sim_core to to use the improved cycle.