Refacted actions between base_object tick and apply.
Tasks #82 and #83 : Refactored tasks so that tick() accumulates and
applies changes to velocity, rotation and mass. apply() now updates
location and attitude based on the the velocities, allowing forward
looking collision checking to happen in between those steps.
Correctness was verified by the unit test, which only required a
change for torque application, as we were using an improper short-cut
before; we are now using a pre-effector as we should have been from
the beginning.