sim_core collision check fix updates complete.
Task #84 : All changes required to make sim_core work with the
collision fix are in place and unit tested. The earlier problem was
due to the unit test not making sure that the sim_core has finished
shutdown before checking the test conditions; that has been fixed
now. sim_core changes are complete and ready to use.