Corrected bad base_object calls in bot_mk1.
Task #78 : Due to the collision fix effort, several signatures in
base_object changed, breaking bot_mk1. Corrected and unit tested.
Discovered a slight alteration in cloning results due to a change
in tick v. apply responsibilities, this was evaluated, found to be
acceptable and the unit test corrected to accept the output.
Compliled, unit tested and ready to go.