Checkpoint on the rail_gun unit test development.
Task #33 : Added a method to the monitor, wait_for_move which allows
the test program to block and return only the current movement is
complete (or the next one, if no movement is currently active). This
sped up the unit test significantly and simplifies the test code,
particularly for the fire on stable test still to be created.

Compliled and run for a few hundred iterations to make sure it's
working reliably. Still need to complete the the fire on stable
test as well as tests to prove targeting is accurate.