Added null pointer test to gp_sim_message::data()
Issue #28 : Added documentation comments to gp_sim_message. While
I was at it, I noticed that the data() method wrongly assume it would
only be called if a data chunk was present.. a very bad assmption.
Altered the method to check for a null pointer and throw if one
is encountered. Compiled and unit tested.