Altered gp_sim_message::data<T>() to return T instead of T &
Issue #28 : As originally written, gp_sim_message::data<T>() returned
a reference to the data packet as T. This is dangerous as you could end
up pointing a released memory area. This revision has it return a copy
of the data packet as T, so the scope is fully controlled by the
caller. Compiled and unit tested.