Issue #39 : Moved bot_interfaces up in the build list.
This was needed because I have determimed that it would be better
if the individual modules handled their command interface. Interfaces
for that were not available at the time time I orignally wrote the
effector and sensor modules.

I'll be moving the command/response processing for each module down
to the individual modules this week as part of the radar integration
issue.