Altered the signiture of the "commandable" interface.
Issue #39 : While integrating the new interface for diff_steer into
bot_mk1, I realized that commandable::command() needed to return a
bool to indicate if it handled the command. This allows the caller
to know if a particular module handled the request or if it needs to
pass it on to the next module. Changes are completed, compiled and
unit tested and ready to go.