Altered bot_mk1 to work with the new diff_steer commandable interface.
Issue #39 : altered bot_mk1::msg_router() to use the new interface to
talk to diff_steer. The unit test required no changes and works properly.
It does test functionality not currently implemented in bot_mk1, as
diff_steer message processing is now in that module and should not require
detailed testing here. Removing the unneeded tests is a task for a
later time.