Added radar to bot_mk1, but no command response yet.
Task #38 : After some linking trouble, a radar_mk1 has been added
to the bot_mk1, but not integrated to the command structure yet.
This code compiles and passes the existing unit tests. More work
is needed to add radar to the BCP command chain and to update
the unit test for the new functionality.