Task #41 : Final (for now) refinements to the target BCP
* Added "zonelimit" configuration parameter, to limit the speed
while in zone transit (cruse) mode.

* Changed from a "stop, calc, and turn" nav model to doing the
calc and turn on the run. This is both faster and keeps us closer
to the zone because we don't continue to overrun while stopping.

* There is no braking used at all in this BCP... skid losses while
turning combined with zone speed limit are sufficient to keep it
reasonably controlled.

This has been tested pretty extensively. It produces reasonable
behavior even when using some pathologically stupid configurations.
Unit tested and ready to go.