Task #41 : Working, if not very efficient kamikaze bot.
This version successfully tracks and seeks the first contact found.
It does no predictive work and is pretty sloppy on steering, so it
has a very hard time hitting even a slowly moving target, but it
does approach to close range and try to hit, even if it keeps missing.
It's enough to prove the concepts, which was the primary goal of this
exercise. Unit tested and ready to run.