Task #51 : Created unique simulation run ID.
Added a method to simengine.cpp which creates a key, unique to the
second for each simulation start, and added that to the document for
the quanta collection.  Still need to find the best way to share the
new id with the rest of the process so it can be used as a common
key to link all persistant data. Compiles and tests clean.