Beta #4 : Made radar_mk1 an effector.
Converted radar_mk1 to a child of abs_effector. Most of it was
actually header changes and unit test updates to use the
base_object command() method. Unit tested and ready to go.

I'm a little concerned that clone maintains the link to the parent
object, but that should never be a problem unless the command
"radar contacts" is passed in. I'll find a way to clean it out
later.

Unit tested and ready to go.