Beta #4 : Changes to bot_mk1 to use the altered base_object, effectors.
Minmal changes to bot_mk1 and unit test. Method msg_router() was
simplified to use base_object::command() to talk to the effectors.
The only unit test change was to the clone test, which failed becase
radar_mk1 is now an effector and therefore showed in the as_str()
output. Changed the test string to match the new expected output.

Unit tested and ready to go.