Beta #2 : radar_mk1 now has a no-argument constructor.
In line with standardizing effector constructors to make them plugable,
radar_mk1 had it's internal reference to the parent object removed and
it's tick() method (previously a NOP) was changed to copy the needed
data from the object to local storage.

Unit tested and ready to go. Changes will be needed to bot_mk1 due
to the constructor signature change.