Beta #14 : Added a ping command to bot_mk1.
Testing with an improved version target_2 I found it could run run over
60 seconds without sending a command since autol was on. This resulted
the sim engine disconnecting it assuming the bot was hung. To address
this I added a new BCP command "bot ping", which does nothing but return
"READY" to the BCP. This commit is the improved version of target_2 and
bot_mk1 change to impliment the ping command. Tested and ready to go.