Beta #18 : Added lock() and unlock methods to base_object.
base_object lock() and unlock() are NOP virtual methods to provide
a method for sim_core to block local processing in objects during
game turn calculation. It's a NOP by default so as to avoid the cost
of locking and unlocking a mutex unnecessarily. Children of
base_object are responsible for overriding as needed to prevent
inappropriate processing during game turn calculation.

Unit tested and ready to go.