Beta #18 : Altered bot_mk1 to respect sim_core lock requests.
Overrode lock and unlock to manage bot_mk1::cooking_lock, and removed
the expliced lock and unlocks which were in the tick and apply methods.
There is still a unique_lock protecting the scope of the msg_router
method, as commands should not be processed while sim_core is busy.

Tested using the old-style target and kamikaze with very good results.
However, we still have a problem with the autol and autor commands. As
a guess it's the fact that the auto commands tend to cause all the BCPs
using them to operate in lockstep, creating massive surges. I'll leave them
active for now, but they they may be a bad idea if I can not find a way to
spread the load a bit.