Beta #1 : Added a simple pre-test to the collision checking.
Added a simple test to the collision checker which avoids the full
check if objects are too far apart to interact in the current quanta.
This resulted in a slight increase in performance, but so little that
we're really lost in weeds a bit.

Conducted tests with collision checking on and completely commented out.
This showed that collision checking is not a major part of the time used
in the quanta calculation, ~10% or less in a 50 object simulation. As
such, there is not much value in pursing this futher. We'll likely be
closing issue #1 after a final review.

Build, unit tested, extensively run tested. This is ready for merge to
master.